requirements.txt
streamlit>=1.37.0
google-genai>=0.3.0
matplotlib>=3.8.0
python-dotenv>=1.0.0



